Tipo incorrecto en el getter.

+++ converter-businessrules-dynamic-package/src/main/java/com/barcelo/businessrules/dynamicpack/decision/impl/DecisionManager.java
@@ -48,7 +48,7 @@ public class DecisionManager {
 		}
 	}
 
-	public long getPollingInterval() {
+	public int getPollingInterval() {
 		return pollingInterval;
 	}
